refactor: 🧑‍💻 move providers to other directory
