+++ src/App.tsx
@@ -5,9 +5,9 @@ import Listing from './components/Listing';
 import { init as initStore, get as getFromStore, set as saveToStore, SAVES_PATH, LAST_SAVE } from './settings';
 import { getVersion } from '@tauri-apps/api/app';
 import ScrollToTopBtn from './components/ScrollToTopBtn';
-import { useDirectoryProvider } from './components/DirectoryProvider';
-import { STS_IDLE, useRawsProvider } from './components/RawsProvider';
-import { useSearchProvider } from './components/SearchProvider';
+import { useDirectoryProvider } from './providers/DirectoryProvider';
+import { STS_IDLE, useRawsProvider } from './providers/RawsProvider';
+import { useSearchProvider } from './providers/SearchProvider';
 import SearchBox from './components/SearchBox';
 import MenuBar from './components/MenuBar';
 import DFDirectoryNotSet from './components/DFDirectoryNotSet';
